Fixing obvious gen server issues and moving to the migrator for a way better design
